got remote hosts working in the nightly check
git-svn-id: file:///Users/gsalam/svn/fastjet/trunk/fastjet-release@1584 9ab96302-e443-4389-8a14-f726967065a6

+++ regression-tests/nightly-check.pl
@@ -1,34 +1,44 @@
 #!/usr/bin/perl -w
 #
-# script to help us perform a nightly check of fastjet
+# Script to help us perform a nightly check of fastjet
 #
-# -mail sends mail, otherwise, verbose output
+# -mail sends mail, otherwise output goes to screen
 #
+# Various other options provide access to internals for running checks
+# on remote hosts. The set of configurations that are run is given in
+# the setups variable below.
 #
 #----------------------------------------------------------------------
-# Items are
+# The script does the following once:
 #   - svn update
 #   - make dist
+#
+# And then the following for each setup
 #   - untar
 #   - configure from separate dir
 #   - make
 #   - make check
-#   - regression-tests/test-all-algs.pl -nev 1000
+#   - regression-tests/test-all-algs.pl -nev [some number]
 #
 #----------------------------------------------------------------------
 # Future options:
-#   - include various other configures (with/without shared libs, cgal, etc.)
-#   - include different compilers
-#     [e.g. /ada1/lpthe/cacciari/local/bin/g++-4.4]
-#   - and try it out on macs too?
+#   - include an "executive summary" either at the end (screen)
+#     or the beginning (mail), at least when things are OK.
 #
-# What should the architecture be? Call this program with options?
-# This program calls others if you ask for specific cases?
+#   - include info on svn revision, and directory status?
 #
 #----------------------------------------------------------------------
 # Reminder notes:
 #
 # NB: $? is command status (non-zero with error)
+#
+# -------- model crontab file (on toth)--------------------------------
+# # select the default shell (to get all paths, etc., e.g. for CGAL)
+# SHELL=/bin/zsh
+#
+# # at 5.34 every morning run the fastjet tests;
+# #
+# 34 05 * * * cd $HOME/work/jets/fjr-branches/fastjet-trunk-nightly-tests ; regression-tests/nightly-check.pl -mail
 
 use Cwd;
 use English;
@@ -44,73 +54,159 @@ $mailAddr='salam@lpthe.jussieu.fr cacciari@lpthe.jussieu.fr gsoyez@quark.phy.bnl
 # $nevTestAll=1000;
 
 @setups = ();
-# for each setup we put in the config options, the special link-time flags, and the number of events
-push @setups, ["", "", 10]; # out of the box
-push @setups, ["--enable-allcxxplugins --enable-cgal --with-cgaldir=".$ENV{CGAL_DIR}, "", 1000]; # with CGAL & all plugins
-push @setups, ["--enable-allcxxplugins --enable-shared --disable-static", "--runpath", 10]; # with dynlibs
-push @setups, ["--enable-allcxxplugins --enable-shared", "--shared=no", 10]; # with static libs even though shared are built
+# for each setup we put the host ("" is current host), the config
+# options, the special link-time flags, and the number of events
+#
+# The things we want to test are:
+#
+# - out of the box compilation on linux
+# - the same on a mac
+# - a full set of algs on toth, mac, a 64 bit machine, gcc 4.4
+# - shared/static issues (depending on current defaults)
+# - cgal
+# - at least one run with 10^3 events
+
+
+
+push @setups, ["karnak","--enable-allcxxplugins CC=/usr/local/bin/gcc-4.4 CXX=/usr/local/bin/g++-4.4", "", 10]; # full set with gcc 4.4 
+
+# $cxx = "g++";
+# # special compilers are deduced from the configure flag
+# if ($setups[0][1]=~ /CXX=([^\s]+)/) { $cxx = $1; }
+# print "$cxx\n";exit;
+
+# push @setups, ["","", "", 10]; # out of the box
+# push @setups, ["","--enable-allcxxplugins --enable-cgal --with-cgaldir=".$ENV{CGAL_DIR}, "", 1000]; # with CGAL & all plugins
+# push @setups, ["","--enable-allcxxplugins --enable-shared --disable-static", "--runpath", 10]; # with dynlibs
+# push @setups, ["","--enable-allcxxplugins --enable-shared", "--shared=no", 10]; # with static libs even though shared are built
+# push @setups, ["zetes", "", "", 10]; # out of the box on zetes (SLC4, gcc 3.4.6, 64 bit)
+# push @setups, ["karnak","", "", 10]; # out of the box on karnak (OS X 10.5)
+# push @setups, ["karnak","--enable-allcxxplugins --enable-shared", "", 10]; # full monty on karnak
+# push @setups, ["hercule","--enable-allcxxplugins", "", 10]; # hercule: standard machine, 64 bits
+
+
+
+# /Volumes/Lacie/cacciari/local/bin/g++-4.4
 
 
 # process command-line
 $mail=0;
+$tmpDir="";
+$remote=0;
+$command=$0;
+$origDir=getcwd();
+$tarName="";
 while ($arg = shift @ARGV) {
-  if ($arg eq "-mail") {$mail = 1;}
+  if ($arg eq "-mail")   {$mail = 1;}
+  elsif ($arg eq "-remote") {$tmpDir  = shift @ARGV; $remote=1;}
+  elsif ($arg eq "-tar")    {$tarName = shift @ARGV;}
+  elsif ($arg eq "-orig")   {
+    $origDir = shift @ARGV;
+  }
   else {die "Unrecognized argument: $arg";}
 }
 
-$origDir=getcwd();
+# for some remote hosts, need to remove leading "/misc?" from directory name
+$origDir=~ s/^\/misc//; 
+
 $fail="";
 $failDetails="";
 $allMessages = "";
-$tmpDir = "";
-$verbose = !$mail;
+$verbose = ! ($mail || $remote);
 #$tarName="fastjet-2.4-devel.tar.gz"; # TMP 
+
+$uname = `uname -a`; chomp $uname;
+&message("* running on $uname\n");
 
 MAIN: while (1) {
 
-  #--- make tmpDir -------------------------------------------------------
-  #$tmpDir = "$origDir/tmp-nightly";
-  $tmpDir = "$origDir/tmp-".$$;
-  $uname = `uname -a`; chomp $uname;
-  &message("* running on $uname\n");
-  &message("* making tmp directory $tmpDir\n");
-  if (-e $tmpDir || ! (mkdir $tmpDir)) {
-    $fail = "* creating tmp directory";
-    $failDetails = "$tmpDir already exists or could not be created; stopping";
-    $tmpDir = "";
-    last;
-  }
+  if (!$tmpDir) {
+    #--- make tmpDir -------------------------------------------------------
+    #$tmpDir = "$origDir/tmp-nightly";
+    $tmpDir = "$origDir/tmp-".$$;
+    &message("* making tmp directory $tmpDir\n");
+    if (-e $tmpDir || ! (mkdir $tmpDir)) {
+      $fail = "* creating tmp directory";
+      $failDetails = "$tmpDir already exists or could not be created; stopping";
+      $tmpDir = "";
+      last;
+    }
 
-  #--- svn update --------------------------------------------------------
-  &message("* running svn update\n");
-  $svnup=`svn update 2>&1`;
-  if ($svnup =~ /external .. revision [0-9]/i && 
-      ($svnup =~ /^At revision [0-9]/m || $svnup =~ /^Updated to revision [0-9]/m) &&
-      $svnup !~ /conflict/i) {
-    # all is OK, do nothing
+    #--- svn update --------------------------------------------------------
+    &message("* running svn update\n");
+    $svnup=`svn update 2>&1`;
+    if ($svnup =~ /external .. revision [0-9]/i && 
+        ($svnup =~ /^At revision [0-9]/m || $svnup =~ /^Updated to revision [0-9]/m) &&
+        $svnup !~ /conflict/i) {
+      # all is OK, do nothing
+    } else {
+      $fail = "svn update";
+      $failDetails = $svnup;
+      last;
+    }
+
+    #--- make dist ------------------------------------------------------
+    &message("* running make dist");
+    $makedist=`make dist 2>&1`;
+    if ($makedist =~ / error[: ]/i || $makedist !~ />(.*?tar.gz)/) {
+      $fail = "make dist";
+      $failDetails = $makedist;
+      &message("\n");
+      last;
+    } else {
+      $tarName = $1;
+      &message(" -> $tarName\n");
+    }
+
+    # now run the rest, either remotely, or from setups array, or from a setup file
+    for ($i = 0; $i <= $#setups; $i++) {
+      if ($setups[$i][0]) {
+        # run test on a remote host 
+        &message("* transferring execution to remote host $setups[$i][0]\n");
+
+        # first set up a file on remote host with the info of interest
+        open(SETUP, "> $tmpDir/setup") || die "Could not write to $tmpDir/setup";
+        for ($j=1; $j <=3; $j++) {print SETUP $setups[$i][$j],"\n";}
+        close SETUP;
+
+        # connect to remote host and run there
+        $ssh=`ssh $setups[$i][0] $origDir/$command -remote $tmpDir -orig $origDir -tar $tarName 2>&1`;
+        $ssh =~ s/^.*in the future\n//mg;   # because karnak's time is wrong
+        $ssh =~ s/^.*slocate.db.*\n//mg;    # because zetes has out of date locate
+        $ssh =~ s/^.*updatedb.*\n//mg; # (which I use on logon...)
+        if ($ssh || $?) {
+          $fail = "connection to $setups[$i][0]";
+          $failDetails = $ssh;
+          last MAIN;
+        }
+
+        # collect the results
+        $results=`cat $tmpDir/messages 2>&1`;
+        if (!$results || $results =~ /Failed/ || $?) {
+          $fail = "execution on remote host";
+          $failDetails = $results;
+          last MAIN;
+        } else {
+          &message($results);
+        }
+
+      } else {
+
+        # run the test locally
+        &build_and_check($setups[$i][1], $setups[$i][2], $setups[$i][3]) || last MAIN;
+
+      }
+    }
   } else {
-    $fail = "svn update";
-    $failDetails = $svnup;
-    last;
+    # remote case, in which tmpDir is already there
+    # read the instructions
+    open(SETUP, "< $tmpDir/setup") || die "failed to read from $tmpDir/setup;";
+    for ($j=0; $j <= 2; $j++) {$setup[$j] = <SETUP>; chomp($setup[$j]);}
+    close SETUP;
+    # execute them
+    &build_and_check($setup[0], $setup[1], $setup[2]) || last MAIN;
   }
 
-  #--- make dist ------------------------------------------------------
-  &message("* running make dist");
-  $makedist=`make dist 2>&1`;
-  if ($makedist =~ / error[: ]/i || $makedist !~ />(.*?tar.gz)/) {
-    $fail = "make dist";
-    $failDetails = $makedist;
-    &message("\n");
-    last;
-  } else {
-    $tarName = $1;
-    &message(" -> $tarName\n");
-  }
-
-  # now run the rest
-  for ($i = 0; $i <= $#setups; $i++) {
-    &build_and_check($setups[$i][0], $setups[$i][1], $setups[$i][2]) || last MAIN;
-  }
   #&build_and_check($configOpts, $fjlibOpts, $nevTestAll) || last;
 
   # now just exit
@@ -120,29 +216,34 @@ MAIN: while (1) {
 
 #-- mention where failure might arise
 if ($fail) {
-  &message("Failed on $fail\n\nDetailed message is:\n------------------");
+  &message("Failed on $fail\n\nDetailed message is:\n------------------\n");
   &message($failDetails);
+  &message("\n--------------- END OF FAILURE MESSAGE ---------------\n");
   $mailSubject='fastjet nightly: FAILED on '.$fail;
-} else {
+} elsif (!$remote) {
   &message("\nAll tests passed\n");
   # try to get more info about test results
-  @unavail = split("unavailable",$testall);
-  @areOK   = split("OK",$testall);
+  @unavail = split("unavailable",$allMessages);
+  @areOK   = split("OK",$allMessages);
   $mailSubject='fastjet nightly: '.sprintf("%d",$#areOK).' OK';
   if ($#unavail >= 0) {$mailSubject .= ", ".sprintf("%d",$#unavail)." NA"}
 }
 
 # clean up
-if ($tmpDir && !$fail) { 
+if ($tmpDir && !$fail && !$remote) { 
   &message("* removing $tmpDir\n");
   system("rm -rf $tmpDir")
 };
 
-# send mail if relevant
+# send mail if relevant, or deposit a message for the program that called us
 if ($mail) {
   open (MAIL, "|mail -s '$mailSubject' $mailAddr") || die "could not open pipe for mail message";
   print MAIL $allMessages;
   close MAIL;
+} elsif ($remote) {
+  open (MSG, "> $tmpDir/messages") || die "Remote host could not write to $tmpDir/messages";
+  print MSG $allMessages;
+  close MSG;
 }
 
 
@@ -167,6 +268,13 @@ sub message ($) {
 sub build_and_check($$$) {
   my ($config,$link,$nev) = @_;
 
+  # $cxx = "g++";
+  # # special compilers are deduced from the configure flag
+  # if ($config =~ /CXX=([^\s]+)/) { $cxx = $1; }
+  # &message("* compiling fastjet_timing_plugins externally (with $cxx, fastjet-config ... $link)\n");
+
+  chdir $tmpDir;
+
   #--- clean up from previous invocation --
   if (-e "build/") {
     &message("\n* removing everything from the tmp dir\n");
@@ -174,8 +282,7 @@ sub build_and_check($$$) {
   }
 
   #--- untar -----------------
-  &message("* untarring in tmp dir\n");
-  chdir $tmpDir;
+  &message("* untarring $origDir/$tarName in tmp dir\n");
   $untar=`tar zxvf $origDir/$tarName`;
   if ($?) {
     $fail = "untar";
@@ -188,10 +295,10 @@ sub build_and_check($$$) {
   chdir "build";
   &message("* running configure $config --prefix=$tmpDir/inst\n");
   ($distDir=$tarName) =~ s/.tar.gz//;
-  $config=`../$distDir/configure $config --prefix=$tmpDir/inst 2>&1`;
-  if ($config =~ /error[: ]/i || $?) {
+  $configOut=`../$distDir/configure $config --prefix=$tmpDir/inst 2>&1`;
+  if ($configOut =~ /error[: ]/i || $?) {
     $fail = "configure";
-    $failDetails = $config;
+    $failDetails = $configOut;
     return 0;
   }
 
@@ -224,8 +331,11 @@ sub build_and_check($$$) {
 
   #--- do external compilation -------------------
   chdir "../";
-  &message("* compiling fastjet_timing_plugins externally (with fastjet-config ... $link)\n");
-  $compile=`g++ -O -I$distDir/example $distDir/example/CmdLine.cc $distDir/example/fastjet_timing_plugins.cc  \`inst/bin/fastjet-config --cxxflags --libs --plugins $link\` -o fastjet_timing_plugins 2>&1`;
+  $cxx = "g++";
+  # special compilers are deduced from the configure flag
+  if ($config=~ /CXX=([^\s]+)/) { $cxx = $1; }
+  &message("* compiling fastjet_timing_plugins externally (with $cxx, fastjet-config ... $link)\n");
+  $compile=`$cxx -O -I$distDir/example $distDir/example/CmdLine.cc $distDir/example/fastjet_timing_plugins.cc  \`inst/bin/fastjet-config --cxxflags --libs --plugins $link\` -o fastjet_timing_plugins 2>&1`;
   if ($compile =~ /error[: ]/i || $?) {
     $fail = "external compilation";
     $failDetails = $compile;
@@ -237,7 +347,7 @@ sub build_and_check($$$) {
   # use the original test-all-algs.pl prog, since it isn't distributed
   # in the tarball
   $testall=`../regression-tests/test-all-algs.pl -nev $nev`;
-  if ($testall =~ /\sBAD/i || $?) {
+  if ($testall =~ /\sBAD/i || $testall !~ /OK/ || $?) {
     $fail = "testing all algs";
     $failDetails = $testall;
     return 0;
